+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.22.0 [unreleased]
 
+### Documentation
+1. [#331](https://github.com/influxdata/influxdb-client-python/pull/331): Add [Migration Guide](MIGRATION_GUIDE.md)
+
 ## 1.21.0 [2021-09-17]
 
 ### Features